+++ stubs/DecodeLabs/Chronos.php
@@ -16,8 +16,8 @@ class Chronos implements Proxy
 {
     use ProxyTrait;
 
-    const VENEER = 'DecodeLabs\\Chronos';
-    const VENEER_TARGET = Inst::class;
+    const Veneer = 'DecodeLabs\\Chronos';
+    const VeneerTarget = Inst::class;
 
     public static Inst $instance;