+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## v0.1.0 (2024-07-24)
 * Built initial Blueprint implementation
 * Added Blueprint Factory
+* Constructed JSON Blueprint schemas