+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Updated Veneer dependency and Stub
+
 ## v0.1.0 (2024-07-24)
 * Built initial Blueprint implementation
 * Added Blueprint Factory