+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Updated Atlas dependency
 * Updated Veneer dependency and Stub
 
 ## v0.1.0 (2024-07-24)